% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{calc_CDFs}
\alias{calc_CDFs}
\title{calc_CDFs}
\usage{
calc_CDFs(DATA, PARS, COPULA)
}
\arguments{
\item{DATA}{Dataframe containing observed time, second event time 
survival probability}

\item{PARS}{Vector of parameter values}

\item{COPULA}{The copula considered by the user.}
}
\description{
Investigate the nature of the cumulative distribution 
and copula functions for simulated or real data sets
}
