% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_surv.R
\name{calc_mDENS_mSURV}
\alias{calc_mDENS_mSURV}
\title{calc_mDENS_mSURV}
\usage{
calc_mDENS_mSURV(vec_time, run_ANA, rDAT, SEXs, AGEs, YEARs)
}
\arguments{
\item{vec_time}{A numeric vector of times in years to calculate 
survival and density values for.}

\item{run_ANA}{The list object outputted from \code{run_analyses()}}

\item{rDAT}{A reference dataset data.frame containing
required columns \code{year}, \code{age}, \code{qx},
and \code{sex} corresponding to reference year, 
reference age, event's interval hazard, and sex, respectively.}

\item{SEXs}{A character vector of sexes to calculate quantities for.
Values should be \code{'male'} or \code{'female'}. Note that input
values will be subsetted to those included in \code{rDAT}}

\item{AGEs}{An integer vector of ages in years. Note that input 
values will be subsetted to those included in \code{rDAT}.}

\item{YEARs}{An integer vector of calendar years diagnosed. 
Note that input values will be subsetted to those included in 
\code{rDAT}.}
}
\description{
Once the optimization has finished and estimates
obtained, the user may calculate net survival and net
densities, strata-specific survivals and densities, and 
observed survivals and observed densities for a set of times
}
