% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{dexpweibull}
\alias{dexpweibull}
\title{dexpweibull}
\usage{
dexpweibull(x, lambda, alpha, kappa, log = FALSE)
}
\arguments{
\item{x}{Quantile}

\item{lambda}{lambda parameter}

\item{alpha}{alpha parameter}

\item{kappa}{kappa parameter}

\item{log}{Boolean}
}
\description{
PDF of exp-weibull
}
