% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{full_ana_est}
\alias{full_ana_est}
\title{full_ana_est}
\usage{
full_ana_est(RES, COHORT)
}
\arguments{
\item{RES}{The output object from \code{full_ana_opt()}}

\item{COHORT}{A string to specify which dataset or cohort
is being analyzed.}
}
\description{
A comprehensive summary of parameter estimates
	in data.frame and outputted table format.
}
