% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{full_ana_opt}
\alias{full_ana_opt}
\title{full_ana_opt}
\usage{
full_ana_opt(DATA, max_year = 100, param_grid = seq(-3, 5, 0.5), verb)
}
\arguments{
\item{DATA}{A data.frame containing column names \code{time},
\code{delta}, \code{dens_t2}, and \code{surv_t2}.}

\item{max_year}{An integer number of years of survival 
probability to calculate.}

\item{param_grid}{Vector of values spanning possible 
log(alpha1), log(lambda1), and log(kappa1) parameters}

\item{verb}{Boolean, set to TRUE for verbose output.}
}
\description{
A comprehensive optimization over copulas
and predicting survival probabilities.
}
