% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{full_ana_plots}
\alias{full_ana_plots}
\title{full_ana_plots}
\usage{
full_ana_plots(curr_dir, RES, NAME)
}
\arguments{
\item{curr_dir}{The directory where data analyses 
are outputted.}

\item{RES}{The output object from \code{full_ana_opt()}}

\item{NAME}{Character string specifying a subdirectory
of \code{curr_dir} to place plots for a specific 
working dataset.}
}
\description{
Visual summary of grid search, 
survival curves.
}
