% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{full_ana_surv}
\alias{full_ana_surv}
\title{full_ana_surv}
\usage{
full_ana_surv(RES, myYEARS = NULL)
}
\arguments{
\item{RES}{The output object from \code{full_ana_opt()}}

\item{myYEARS}{Defaults to \code{NULL} to list
survivals per year. Otherwise supply an integer 
vector of years of survival probabilies to output}
}
\description{
Summarizes survival probabilities
	in a Latex output and data.frame object
}
