% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{full_sim}
\alias{full_sim}
\title{full_sim}
\usage{
full_sim(
  copula,
  dist1,
  NN,
  theta,
  alpha1,
  lambda1,
  kappa1,
  alpha2,
  lambda2,
  propC,
  RR,
  param_grid,
  upKAPPA,
  gTHRES = 0.08,
  verb,
  ncores = 1
)
}
\arguments{
\item{copula}{A string input, either "Clayton" or "Gumbel"}

\item{dist1}{A string input for distribution of work dataset 
event times, either "weibull" or "expweibull" for Weibull with
two parameters or exponentiated weibull.}

\item{NN}{Positive integer for simulated sample size}

\item{theta}{A positive value corresponding to the event time
dependency depending on the copula. For Clayton, \code{theta} >= 0
and for Gumbel, \code{theta} >= 1.}

\item{alpha1}{First shape parameter for the first event.}

\item{lambda1}{Scale parameter for the first event.}

\item{kappa1}{Second shape parameter for the first
event's exponentiated Weibull.}

\item{alpha2}{Shape parameter for the second/reference event.}

\item{lambda2}{Scale parameter for the second/reference event.}

\item{propC}{A numeric value between 0 and 1 for
proportion of subjects with censored event times}

\item{RR}{Integer number of replicates}

\item{param_grid}{Vector of values spanning possible 
log(alpha1), log(lambda1), log(kappa1), unconstrained 
theta parameters}

\item{upKAPPA}{An integer value taking values 0 or 1. If set 
to 1, the exponentiated Weibull distribution is assumed. Otherwise,
the Weibull distribution is assumed and optimized over.}

\item{gTHRES}{A numeric threshold on the L2 norm of 
the gradient evaluated at the MLE.}

\item{verb}{Boolean, set to TRUE for verbose output.}

\item{ncores}{An integer for the number of threads}
}
\description{
This function performs a full simulation after 
supplying the corresponding copula, sample size, underlying
parameters, number of simulation replicates, and grid of 
values to search over.
}
