% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{opt_replicate}
\alias{opt_replicate}
\title{opt_replicate}
\usage{
opt_replicate(
  REP,
  param_grid,
  theta = NULL,
  upKAPPA,
  ncores = 1,
  gTHRES = 0.08,
  verb
)
}
\arguments{
\item{REP}{An object generated by \code{sim_replicate}.}

\item{param_grid}{Vector of values spanning possible 
log(alpha1), log(lambda1), log(kappa1), unconstrained 
theta parameters}

\item{theta}{A positive value corresponding to the event time
dependency depending on the copula. For Clayton, \code{theta} >= 0
and for Gumbel, \code{theta} >= 1.}

\item{upKAPPA}{An integer value taking values 0 or 1. If set 
to 1, the exponentiated Weibull distribution is assumed. Otherwise,
the Weibull distribution is assumed and optimized over.}

\item{ncores}{An integer for the number of threads}

\item{gTHRES}{A numeric threshold on the L2 norm of 
the gradient evaluated at the MLE.}

\item{verb}{Boolean, set to TRUE for verbose output.}
}
\description{
This function optimizes the dataset generated by 
\code{sim_replicate}. First this function runs a grid search
for optimal initial parameter estimates and then proceeds to
optimize using BFGS.
}
