% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{pexpweibull}
\alias{pexpweibull}
\title{pexpweibull}
\usage{
pexpweibull(q, lambda, alpha, kappa, log.p = FALSE)
}
\arguments{
\item{q}{quantile}

\item{lambda}{lambda parameter}

\item{alpha}{alpha parameter}

\item{kappa}{kappa parameter}

\item{log.p}{Boolean}
}
\description{
CDF of exp-weibull
}
