% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{qexpweibull}
\alias{qexpweibull}
\title{qexpweibull}
\usage{
qexpweibull(p, lambda, alpha, kappa)
}
\arguments{
\item{p}{Probability}

\item{lambda}{lambda parameter}

\item{alpha}{alpha parameter}

\item{kappa}{kappa parameter}
}
\description{
Inv-quantile
}
