% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{sim_replicate}
\alias{sim_replicate}
\title{sim_replicate}
\usage{
sim_replicate(
  copula,
  dist1,
  NN,
  theta,
  alpha1,
  lambda1,
  kappa1,
  alpha2,
  lambda2,
  propC,
  show
)
}
\arguments{
\item{copula}{A string input, either "Clayton" or "Gumbel"}

\item{dist1}{A string input for distribution of work dataset 
event times, either "weibull" or "expweibull" for Weibull with
two parameters or exponentiated weibull.}

\item{NN}{Positive integer for simulated sample size}

\item{theta}{A positive value corresponding to the event time
dependency depending on the copula. For Clayton, \code{theta} >= 0
and for Gumbel, \code{theta} >= 1.}

\item{alpha1}{First shape parameter for the first event.}

\item{lambda1}{Scale parameter for the first event.}

\item{kappa1}{Second shape parameter for the first
event's exponentiated Weibull.}

\item{alpha2}{Shape parameter for the second/reference event.}

\item{lambda2}{Scale parameter for the second/reference event.}

\item{propC}{A numeric value between 0 and 1 for
proportion of subjects with censored event times}

\item{show}{Boolean, set to TRUE for verbose output.}
}
\description{
This function simulates an analysis-ready dataset
with arguments to specify the sample size and copula-related and 
survival-related parameters.
}
