% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa}
\alias{usa}
\title{USA men and women population dataset}
\format{
An object of class \code{data.frame} with 19536 rows and 4 columns.
}
\source{
\href{https://www.mortality.org/}{Mortality Database}
}
\usage{
usa
}
\description{
USA men and women population dataset
}
\keyword{datasets}
