% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_LL}
\alias{plot_LL}
\title{plot_LL}
\usage{
plot_LL(GPROF, GOPT = NULL, nBREAKs = 5, COPULA, HJUST = 1)
}
\arguments{
\item{GPROF}{An outputted data.frame from \code{get_PROFILE()}.}

\item{GOPT}{An outputted data.frame from \code{get_LOCAL_OPTS()}.}

\item{nBREAKs}{A positive integer set to 5 by default to specify
the number of breaks in axes ticks.}

\item{COPULA}{A string input, either "Clayton" or "Gumbel"}

\item{HJUST}{A numeric value for horizontal alignment of labeled
ggplot points}
}
\description{
Visualize grid of profile log likelihood
}
