% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{refData_match}
\alias{refData_match}
\title{refData_match}
\usage{
refData_match(wDAT, rDAT, ncores = 1)
}
\arguments{
\item{wDAT}{A working dataset data.frame containing 
required columns \code{age}, \code{time}, \code{datediag_yr}, 
and \code{sex} corresponding to age, observed time, 
diagnosis year, and sex (coded 'female'/'male'), respectively.}

\item{rDAT}{A reference dataset data.frame containing
required columns \code{year}, \code{age}, \code{qx},
and \code{sex} corresponding to reference year, 
reference age, event's interval hazard, and sex, respectively.}

\item{ncores}{Integer number of parallel threads to decrease
matching runtime.}
}
\value{
A dataframe containing calculated log-transformed density and 
log-transformed cumulative distribution
}
\description{
This function takes as input
a working dataset of interest and a 
reference dataset.
}
