% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMrs.R
\name{run_analyses}
\alias{run_analyses}
\title{run_analyses}
\usage{
run_analyses(
  DATA,
  THETAs = NULL,
  upKAPPA,
  gTHRES = 0.08,
  copula,
  param_grid = seq(-3, 3, 0.25),
  vec_time,
  ncores = 1,
  show
)
}
\arguments{
\item{DATA}{A data.frame containing column names \code{time},
\code{delta}, \code{dens_t2}, and \code{surv_t2}.}

\item{THETAs}{A vector of theta values to explore and optimize over.}

\item{upKAPPA}{An integer value taking values 0 or 1. If set 
to 1, the exponentiated Weibull distribution is assumed. Otherwise,
the Weibull distribution is assumed and optimized over.}

\item{gTHRES}{A numeric threshold on the L2 norm of 
the gradient evaluated at the MLE.}

\item{copula}{A string input, either "Clayton" or "Gumbel"}

\item{param_grid}{Vector of values spanning possible 
log(alpha1), log(lambda1), and log(kappa1) parameters}

\item{vec_time}{Vector of times to calculate predicted survival
on the same scale as times provided in the \code{DATA} data.frame.}

\item{ncores}{An integer for the number of threads}

\item{show}{Boolean, set to TRUE for verbose output.}
}
\description{
This function performs a full analysis of 
an inputted dataframe. The user may specify one of two 
copulas, a \code{theta} value, a parametric grid to 
search over, and a vector of times for predicting survival.
}
